requirements.txt
Django==4.2.1
django-cors-headers==3.14.0
django-filter==23.1
djangorestframework==3.14.0
Markdown==3.4.3
django-environ==0.10.0
pillow==9.5.0
django-storages==1.13.2
django-ckeditor==6.5.1
psycopg2==2.9.6
